services/recurring-task/pyproject.toml
[project]
name = "recurring-task-service"
version = "1.0.0"
description = "Recurring task automation service for Phase V cloud-native architecture"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "python-dateutil>=2.9.0",  # RFC 5545 recurrence patterns
    "dapr>=1.12.0",
    "dapr-ext-fastapi>=1.12.0",
    "pydantic>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
